cleaned up a buncha shit
